채팅 정리하지 못했던 코드 백업

+++ BookChat/app/src/main/java/com/example/bookchat/data/response/CursorMeta.kt
@@ -16,5 +16,5 @@ data class CursorMeta(
     @SerializedName("first")
     val first :Boolean,
     @SerializedName("nextCursorId")
-    val nextCursorId :Int
+    val nextCursorId :Long
 )
